server/PusoyDosOnline/PusoyDosServer/serializers.py
from django.contrib.auth import get_user_model
from django.contrib.auth.password_validation import validate_password
from rest_framework import serializers
from rest_framework.validators import UniqueValidator


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = get_user_model()
        fields = ['username', 'rating', 'country_code', 'played_games', 'won_games', 'lost_games', 'winstreak']
