server/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
psycopg2==2.9.3
djangorestframework-simplejwt==5.1.0
django-countries==7.3.2